refactor: dto -> record, 디렉토리 변경
