--- src/main/java/com/coniverse/dangjang/domain/auth/dto/Response/LoginResponse.java
@@ -1,20 +1,0 @@
-package com.coniverse.dangjang.domain.auth.dto.Response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-/**
- * 로그인 RESPONSE
- *
- * @author EVE
- * @since 1.0
- */
-@AllArgsConstructor
-@Getter
-public class LoginResponse {
-	private String oauthId;
-	private String nickname;
-	private String accessToken;
-	private String refreshToken;
-	private Long expiresIn;
-}
