+++ src/main/java/com/coniverse/dangjang/domain/auth/dto/request/KakaoLoginRequest.java
@@ -2,22 +2,15 @@ package com.coniverse.dangjang.domain.auth.dto.request;
 
 import com.coniverse.dangjang.domain.auth.dto.OauthProvider;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import jakarta.validation.constraints.NotBlank;
 
 /**
- * 카카오 로그인 parm
+ * 카카오 로그인 dto
  *
- * @author EVE
- * @since 1.0
+ * @author EVE, TEO
+ * @since 1.0.0
  */
-@Getter
-@Setter
-@NoArgsConstructor
-public class KakaoLoginRequest implements OauthLoginRequest {
-	private String accessToken;
-
+public record KakaoLoginRequest(@NotBlank(message = "") String accessToken) implements OauthLoginRequest {
 	@Override
 	public OauthProvider getOauthProvider() {
 		return OauthProvider.KAKAO;
